hash: change hash_value overloading

+++ libs/exo/include/exo/collections/handle.h
@@ -6,7 +6,8 @@ namespace exo
 /// --- Handle type (Typed index that can be invalid)
 usize _hash_handle(u32 index, u32 gen);
 
-template <typename T> struct Handle
+template <typename T>
+struct Handle
 {
 	static constexpr Handle invalid() { return Handle(); }
 	constexpr Handle() : Handle(u32_invalid, u32_invalid) {}
@@ -23,10 +24,13 @@ private:
 	u32 index;
 	u32 gen;
 
-	template <typename T> friend struct Pool;
-	template <typename T> friend struct PoolIterator;
-	template <typename T> friend struct ConstPoolIterator;
-	friend inline usize hash_value(const Handle<T> &h) { return _hash_handle(h.index, h.gen); }
+	template <typename T>
+	friend struct Pool;
+	template <typename T>
+	friend struct PoolIterator;
+	template <typename T>
+	friend struct ConstPoolIterator;
+	friend inline u64 hash_value(const Handle<T> &h) { return _hash_handle(h.index, h.gen); }
 };
 
 } // namespace exo
